move Store class and some helper functions to zarr._storage.store
update version in Store docstring
